+++ src/lib/compiler/physical_query_plan/exchange/abstract_exchange_strategy.cpp
@@ -1,0 +1,10 @@
+#include "abstract_exchange_strategy.hpp"
+
+namespace skyrise {
+
+AbstractExchangeStrategy::AbstractExchangeStrategy(const ExchangeStrategyType type)
+    : type_(type) {}
+
+ExchangeStrategyType AbstractExchangeStrategy::Type() const { return type_; }
+
+}  // namespace skyrise
